loader: Override layer support and settings
Add support for an override layer in the loader. This allows layers
to be set by an outside application.

Revamp the logic to detect manifest files in the various paths that
we allow them to be placed. This code significantly rewrites
loader_get_manifest_files.

Also, add the logic to find the layer settings file using the loader.
That way we can find the proper path once and provide it to all
layers. This also allows us to place the settings file in a global
directory so we can override it with the pending layer manager GUI.

Change-Id: I8abf558864b66eb71ee026ca559b0126cf2fa4e9